QStorageInfo: change private blockSize member to int
It's what the public API returns. There's little reason to use ulong,
despite that being the type that struct statvfs uses (on Windows, that
was 32-bit anyway).

Saves 8 bytes on the size of the QStorageInfoPrivate class.

+++ src/corelib/io/qstorageinfo_p.h
@@ -96,7 +96,7 @@ public:
     qint64 bytesTotal = -1;
     qint64 bytesFree = -1;
     qint64 bytesAvailable = -1;
-    ulong blockSize = ulong(-1);
+    int blockSize = -1;
 
     bool readOnly = false;
     bool ready = false;
